chore: ui improvements
- Added icon to panel size slider
- Added description to panel size slider
- Moved size module to separate file
- Added icons to remaining toggles
- Added entries to fluent files
- Added four new icons to the app
