+++ i18n/ja/cosmic_ext_tweaks.ftl
@@ -25,6 +25,8 @@ show-more = さらに表示
 
 show-panel = パネルを表示
 force-icon-buttons-in-panel = 常にパネルでボタンを表示
+size = Size
+size-description = The height of the panel and applets
 
 padding = 余白
 padding-description = 内容と枠線の間の空間です。
