+++ i18n/sr-Latn/cosmic_ext_tweaks.ftl
@@ -28,6 +28,8 @@ find-themes = Pronađi teme
 # Panel only
 show-panel = Prikaz panela
 force-icon-buttons-in-panel = Nateraj prikaz dugmadi sa ikonicama na panelu
+size = Size
+size-description = The height of the panel and applets
 
 padding = Punjenje
 padding-description = Punjenje je prostor između sadržaja i ivica doka ili panela.
